# By EpokK (26) and Bart Butler (2)
# Via EpokK (14) and Richard (1)
* 'v3' of https://github.com/ProtonMail/Angular: (28 commits)
  we don't use `proton.advSearch`
  Fix #1837
  Fix #1842
  get Time
  Before to store the Time value we check if this parameter is defined. Can solve probably many issue with the Time
  Remove preloading for sent and inbox folder
  Fix #1843
  Correct loading for Stripe JS file
  Fix #1835
  Fix #1828
  Fix #1831
  Fix #1833
  add warning notification style
  Work for #1829
  Change cacheContext detection
  Fix #1829
  Fix #1818
  Move Stripe JS to the dashboard page
  Improve longReadableTime
  use BASE_SIZE constant
  ...

Conflicts:
	src/app/templates/partials/message.tpl.html

+++ src/app/services/auth.js
@@ -18,6 +18,7 @@ angular.module("proton.authentication", [
     pmcw,
     url
 ) {
+    var keys = {}; // Store decrypted keys
 
     // PRIVATE FUNCTIONS
     var auth = {
@@ -38,6 +39,7 @@ angular.module("proton.authentication", [
             } else {
                 // we need the old stuff for now
                 $log.debug('setAuthHeaders:2');
+                $http.defaults.headers.common["x-pm-session"] = undefined;
                 $http.defaults.headers.common.Authorization = "Bearer " + auth.data.AccessToken;
                 $http.defaults.headers.common["x-pm-uid"] = auth.data.Uid;
             }
@@ -65,20 +67,30 @@ angular.module("proton.authentication", [
                             ]).then(
                                 function(result) {
                                     if(angular.isDefined(result[0].data) && result[0].data.Code === 1000 && angular.isDefined(result[1].data) && result[1].data.Code === 1000) {
-                                        user.Role = 1; // Override Role value for the release
+                                        var mailboxPassword = api.getPassword();
+                                        var promises = [];
+
                                         user.Contacts = result[0].data.Contacts;
                                         user.Labels = result[1].data.Labels;
 
-                                        // Set the rows / columns mode
-                                        if (angular.isDefined(user.ViewLayout)) {
-                                            if (user.ViewLayout === 0) {
-                                                $rootScope.layoutMode = 'columns';
-                                            } else {
-                                                $rootScope.layoutMode = 'rows';
-                                            }
-                                        }
+                                        // All private keys are decrypted with the mailbox password and stored in a `keys` array
+                                        _.each(user.Addresses, function(address) { // For each addresses
+                                            _.each(address.Keys, function(key, index) { // For each keys
+                                                promises.push(pmcw.decryptPrivateKey(key.PrivateKey, mailboxPassword).then(function(package) { // Decrypt private key with the mailbox password
+                                                    keys[address.ID] = keys[address.ID] || []; // Initialize array for the address
+                                                    keys[address.ID].push(package); // Add key package
+                                                }, function(error) {
+                                                    // If the primary (first) key for address does not decrypt, display error.
+                                                    if(index === 0) {
+                                                        // TODO display error
+                                                    }
+                                                }));
+                                            });
+                                        });
 
-                                        deferred.resolve(user);
+                                        $q.all(promises).then(function() {
+                                            deferred.resolve(user);
+                                        });
                                     } else if(angular.isDefined(result[0].data) && result[0].data.Error) {
                                         deferred.reject({message: result[0].data.Error});
                                         api.logout(true);
@@ -155,6 +167,9 @@ angular.module("proton.authentication", [
             pmcw.setMailboxPassword(pwd);
         },
 
+        /**
+         * Return the mailbox password stored in the session storage
+         */
         getPassword: function() {
             var pwd = window.sessionStorage[CONSTANTS.MAILBOX_PASSWORD_KEY];
 
@@ -208,9 +223,13 @@ angular.module("proton.authentication", [
             }.bind(this));
         },
 
-        getPrivateKeys: function() {
-            // authentication.user.
-            // TODO need Bart
+        /**
+         * Return the private keys available for a specific address ID
+         * @param {String} addressID
+         * @return {Array}
+         */
+        getPrivateKeys: function(addressID) {
+            return keys[addressID];
         },
 
         getRefreshCookie: function() {
@@ -225,6 +244,9 @@ angular.module("proton.authentication", [
                         window.sessionStorage.setItem(CONSTANTS.OAUTH_KEY+':SessionToken', pmcw.encode_base64(response.data.SessionToken));
                         $log.debug('after',$http.defaults.headers.common['x-pm-session']);
                         $rootScope.doRefresh = true;
+                    }
+                    else {
+                        return $q.reject(response.data.Error);
                     }
                 },
                 function(err) {
@@ -402,7 +424,8 @@ angular.module("proton.authentication", [
          * Removes all connection data
          * @param {Boolean} redirect - Redirect at the end the user to the login page
          */
-        logout: function(redirect) {
+        logout: function(redirect, call_api) {
+            call_api = angular.isDefined(call_api) ? call_api : true;
             var sessionToken = window.sessionStorage[CONSTANTS.OAUTH_KEY+":SessionToken"];
             var uid = window.sessionStorage[CONSTANTS.OAUTH_KEY+":Uid"];
             var process = function() {
@@ -415,7 +438,7 @@ angular.module("proton.authentication", [
 
             $rootScope.loggingOut = true;
 
-            if(angular.isDefined(sessionToken) || angular.isDefined(uid)) {
+            if( call_api && ( angular.isDefined(sessionToken) || angular.isDefined(uid) ) ) {
                 $http.delete(url.get() + "/auth").then(process, process);
             } else {
                 process();
@@ -423,6 +446,10 @@ angular.module("proton.authentication", [
         },
 
         clearData: function() {
+            // Reset $http server
+            $http.defaults.headers.common["x-pm-session"] = undefined;
+            $http.defaults.headers.common.Authorization = undefined;
+            $http.defaults.headers.common["x-pm-uid"] = undefined;
             // Completely clear sessionstorage
             window.sessionStorage.clear();
             // Delete data key
